Extract lambda rewriting code into a separate class
It is a preparation for adding more logic there.
