+++ subprojects/core/src/main/java/org/gradle/internal/classpath/InstrumentingTransformer.java
@@ -28,7 +28,6 @@ import org.gradle.internal.lazy.Lazy;
 import org.gradle.model.internal.asm.MethodVisitorScope;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.Handle;
-import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Type;
 import org.objectweb.asm.tree.ClassNode;
@@ -36,13 +35,9 @@ import org.objectweb.asm.tree.MethodNode;
 
 import java.io.File;
 import java.lang.invoke.CallSite;
-import java.lang.invoke.LambdaMetafactory;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
-import java.lang.invoke.SerializedLambda;
 import java.lang.reflect.InvocationTargetException;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -51,7 +46,8 @@ import java.util.Properties;
 import java.util.function.Supplier;
 
 import static org.gradle.internal.classanalysis.AsmConstants.ASM_LEVEL;
-import static org.objectweb.asm.Opcodes.ACC_INTERFACE;
+import static org.gradle.internal.classpath.CommonTypes.NO_EXCEPTIONS;
+import static org.gradle.internal.classpath.CommonTypes.STRING_TYPE;
 import static org.objectweb.asm.Opcodes.ACC_PRIVATE;
 import static org.objectweb.asm.Opcodes.ACC_STATIC;
 import static org.objectweb.asm.Opcodes.ACC_SYNTHETIC;
@@ -72,16 +68,12 @@ class InstrumentingTransformer implements CachedClasspathTransformer.Transform {
     private static final int DECORATION_FORMAT = 29;
 
     private static final Type SYSTEM_TYPE = getType(System.class);
-    private static final Type STRING_TYPE = getType(String.class);
     private static final Type INTEGER_TYPE = getType(Integer.class);
     private static final Type INSTRUMENTED_TYPE = getType(Instrumented.class);
-    private static final Type OBJECT_TYPE = getType(Object.class);
-    private static final Type SERIALIZED_LAMBDA_TYPE = getType(SerializedLambda.class);
     private static final Type LONG_TYPE = getType(Long.class);
     private static final Type BOOLEAN_TYPE = getType(Boolean.class);
     public static final Type PROPERTIES_TYPE = getType(Properties.class);
 
-    private static final String RETURN_STRING = getMethodDescriptor(STRING_TYPE);
     private static final String RETURN_STRING_FROM_STRING = getMethodDescriptor(STRING_TYPE, STRING_TYPE);
     private static final String RETURN_STRING_FROM_STRING_STRING = getMethodDescriptor(STRING_TYPE, STRING_TYPE, STRING_TYPE);
     private static final String RETURN_STRING_FROM_STRING_STRING_STRING = getMethodDescriptor(STRING_TYPE, STRING_TYPE, STRING_TYPE, STRING_TYPE);
@@ -99,15 +91,12 @@ class InstrumentingTransformer implements CachedClasspathTransformer.Transform {
     private static final String RETURN_LONG_FROM_STRING_LONG_STRING = getMethodDescriptor(LONG_TYPE, STRING_TYPE, LONG_TYPE, STRING_TYPE);
     private static final String RETURN_PRIMITIVE_BOOLEAN_FROM_STRING = getMethodDescriptor(Type.BOOLEAN_TYPE, STRING_TYPE);
     private static final String RETURN_PRIMITIVE_BOOLEAN_FROM_STRING_STRING = getMethodDescriptor(Type.BOOLEAN_TYPE, STRING_TYPE, STRING_TYPE);
-    private static final String RETURN_OBJECT_FROM_INT = getMethodDescriptor(OBJECT_TYPE, Type.INT_TYPE);
-    private static final String RETURN_BOOLEAN_FROM_OBJECT = getMethodDescriptor(Type.BOOLEAN_TYPE, OBJECT_TYPE);
     private static final String RETURN_PROPERTIES = getMethodDescriptor(PROPERTIES_TYPE);
     private static final String RETURN_PROPERTIES_FROM_STRING = getMethodDescriptor(PROPERTIES_TYPE, STRING_TYPE);
     private static final String RETURN_VOID_FROM_PROPERTIES = getMethodDescriptor(Type.VOID_TYPE, PROPERTIES_TYPE);
     private static final String RETURN_VOID_FROM_PROPERTIES_STRING = getMethodDescriptor(Type.VOID_TYPE, PROPERTIES_TYPE, STRING_TYPE);
     private static final String RETURN_CALL_SITE_ARRAY = getMethodDescriptor(getType(CallSiteArray.class));
     private static final String RETURN_VOID_FROM_CALL_SITE_ARRAY = getMethodDescriptor(Type.VOID_TYPE, getType(CallSiteArray.class));
-    private static final String RETURN_OBJECT_FROM_SERIALIZED_LAMBDA = getMethodDescriptor(OBJECT_TYPE, SERIALIZED_LAMBDA_TYPE);
     private static final String RETURN_MAP = getMethodDescriptor(getType(Map.class));
     private static final String RETURN_MAP_FROM_STRING = getMethodDescriptor(getType(Map.class), STRING_TYPE);
 
@@ -168,9 +157,6 @@ class InstrumentingTransformer implements CachedClasspathTransformer.Transform {
     private static final String RETURN_PROCESS_FROM_LIST_LIST_FILE = getMethodDescriptor(PROCESS_TYPE, LIST_TYPE, LIST_TYPE, FILE_TYPE);
     private static final String RETURN_PROCESS_FROM_LIST_LIST_FILE_STRING = getMethodDescriptor(PROCESS_TYPE, LIST_TYPE, LIST_TYPE, FILE_TYPE, STRING_TYPE);
 
-    private static final String LAMBDA_METAFACTORY_TYPE = getType(LambdaMetafactory.class).getInternalName();
-    private static final String LAMBDA_METAFACTORY_METHOD_DESCRIPTOR = getMethodDescriptor(getType(CallSite.class), getType(MethodHandles.Lookup.class), STRING_TYPE, getType(MethodType.class), getType(Object[].class));
-
     private static final String GROOVY_INDY_INTERFACE_TYPE = getType(IndyInterface.class).getInternalName();
 
     @SuppressWarnings("deprecation")
@@ -179,10 +165,6 @@ class InstrumentingTransformer implements CachedClasspathTransformer.Transform {
 
     private static final String INSTRUMENTED_CALL_SITE_METHOD = "$instrumentedCallSiteArray";
     private static final String CREATE_CALL_SITE_ARRAY_METHOD = "$createCallSiteArray";
-    private static final String DESERIALIZE_LAMBDA = "$deserializeLambda$";
-    private static final String RENAMED_DESERIALIZE_LAMBDA = "$renamedDeserializeLambda$";
-
-    private static final String[] NO_EXCEPTIONS = new String[0];
 
     @Override
     public void applyConfigurationTo(Hasher hasher) {
@@ -192,40 +174,29 @@ class InstrumentingTransformer implements CachedClasspathTransformer.Transform {
 
     @Override
     public Pair<RelativePath, ClassVisitor> apply(ClasspathEntryVisitor.Entry entry, ClassVisitor visitor, ClassData classData) {
-        return Pair.of(entry.getPath(), new InstrumentingVisitor(new InstrumentingBackwardsCompatibilityVisitor(visitor), classData::readClassAsNode));
+        return Pair.of(entry.getPath(), new InstrumentingVisitor(new LambdaSerializationTransformer(new InstrumentingBackwardsCompatibilityVisitor(visitor)), classData::readClassAsNode));
     }
 
     private static class InstrumentingVisitor extends ClassVisitor {
         String className;
         private final Supplier<ClassNode> classAsNode;
-        private final List<LambdaFactoryDetails> lambdaFactories = new ArrayList<>();
         private boolean hasGroovyCallSites;
-        private boolean hasDeserializeLambda;
-        private boolean isInterface;
 
         public InstrumentingVisitor(ClassVisitor visitor, Supplier<ClassNode> classAsNode) {
             super(ASM_LEVEL, visitor);
             this.classAsNode = classAsNode;
         }
 
-        public void addSerializedLambda(LambdaFactoryDetails lambdaFactoryDetails) {
-            lambdaFactories.add(lambdaFactoryDetails);
-        }
-
         @Override
         public void visit(int version, int access, String name, String signature, String superName, String[] interfaces) {
             super.visit(version, access, name, signature, superName, interfaces);
             this.className = name;
-            this.isInterface = (access & ACC_INTERFACE) != 0;
         }
 
         @Override
         public MethodVisitor visitMethod(int access, String name, String descriptor, String signature, String[] exceptions) {
             if (name.equals(CREATE_CALL_SITE_ARRAY_METHOD) && descriptor.equals(RETURN_CALL_SITE_ARRAY)) {
                 hasGroovyCallSites = true;
-            } else if (name.equals(DESERIALIZE_LAMBDA) && descriptor.equals(RETURN_OBJECT_FROM_SERIALIZED_LAMBDA)) {
-                hasDeserializeLambda = true;
-                return super.visitMethod(access, RENAMED_DESERIALIZE_LAMBDA, descriptor, signature, exceptions);
             }
             MethodVisitor methodVisitor = super.visitMethod(access, name, descriptor, signature, exceptions);
             Lazy<MethodNode> asMethodNode = Lazy.unsafe().of(() -> {
@@ -242,59 +213,7 @@ class InstrumentingTransformer implements CachedClasspathTransformer.Transform {
             if (hasGroovyCallSites) {
                 generateCallSiteFactoryMethod();
             }
-            if (!lambdaFactories.isEmpty() || hasDeserializeLambda) {
-                generateLambdaDeserializeMethod();
-            }
             super.visitEnd();
-        }
-
-        private void generateLambdaDeserializeMethod() {
-            new MethodVisitorScope(visitStaticPrivateMethod(DESERIALIZE_LAMBDA, RETURN_OBJECT_FROM_SERIALIZED_LAMBDA)) {{
-                Label next = null;
-                for (LambdaFactoryDetails factory : lambdaFactories) {
-                    if (next != null) {
-                        visitLabel(next);
-                        _F_SAME();
-                    }
-                    next = new Label();
-                    Handle implHandle = (Handle) factory.bootstrapMethodArguments.get(1);
-
-                    _ALOAD(0);
-                    _INVOKEVIRTUAL(SERIALIZED_LAMBDA_TYPE, "getImplMethodName", RETURN_STRING);
-                    _LDC(implHandle.getName());
-                    _INVOKEVIRTUAL(OBJECT_TYPE, "equals", RETURN_BOOLEAN_FROM_OBJECT);
-                    _IFEQ(next);
-
-                    _ALOAD(0);
-                    _INVOKEVIRTUAL(SERIALIZED_LAMBDA_TYPE, "getImplMethodSignature", RETURN_STRING);
-                    _LDC(implHandle.getDesc());
-                    _INVOKEVIRTUAL(OBJECT_TYPE, "equals", RETURN_BOOLEAN_FROM_OBJECT);
-                    _IFEQ(next);
-
-                    Type[] argumentTypes = Type.getArgumentTypes(factory.descriptor);
-                    for (int i = 0; i < argumentTypes.length; i++) {
-                        _ALOAD(0);
-                        _LDC(i);
-                        _INVOKEVIRTUAL(SERIALIZED_LAMBDA_TYPE, "getCapturedArg", RETURN_OBJECT_FROM_INT);
-                        _UNBOX(argumentTypes[i]);
-                    }
-                    _INVOKEDYNAMIC(factory.name, factory.descriptor, factory.bootstrapMethodHandle, factory.bootstrapMethodArguments);
-                    _ARETURN();
-                }
-                if (next != null) {
-                    visitLabel(next);
-                    _F_SAME();
-                }
-                if (hasDeserializeLambda) {
-                    _ALOAD(0);
-                    _INVOKESTATIC(className, RENAMED_DESERIALIZE_LAMBDA, RETURN_OBJECT_FROM_SERIALIZED_LAMBDA, isInterface);
-                } else {
-                    _ACONST_NULL();
-                }
-                _ARETURN();
-                visitMaxs(0, 0);
-                visitEnd();
-            }};
         }
 
         private void generateCallSiteFactoryMethod() {
@@ -530,20 +449,7 @@ class InstrumentingTransformer implements CachedClasspathTransformer.Transform {
 
         @Override
         public void visitInvokeDynamicInsn(String name, String descriptor, Handle bootstrapMethodHandle, Object... bootstrapMethodArguments) {
-            if (bootstrapMethodHandle.getOwner().equals(LAMBDA_METAFACTORY_TYPE) && bootstrapMethodHandle.getName().equals("metafactory")) {
-                Handle altMethod = new Handle(
-                    H_INVOKESTATIC,
-                    LAMBDA_METAFACTORY_TYPE,
-                    "altMetafactory",
-                    LAMBDA_METAFACTORY_METHOD_DESCRIPTOR,
-                    false
-                );
-                List<Object> args = new ArrayList<>(bootstrapMethodArguments.length + 1);
-                Collections.addAll(args, bootstrapMethodArguments);
-                args.add(LambdaMetafactory.FLAG_SERIALIZABLE);
-                super.visitInvokeDynamicInsn(name, descriptor, altMethod, args.toArray());
-                owner.addSerializedLambda(new LambdaFactoryDetails(name, descriptor, altMethod, args));
-            } else if (isGroovyIndyCallsite(bootstrapMethodHandle)) {
+            if (isGroovyIndyCallsite(bootstrapMethodHandle)) {
                 Handle interceptor = new Handle(
                     H_INVOKESTATIC,
                     INSTRUMENTED_TYPE.getInternalName(),
@@ -566,20 +472,6 @@ class InstrumentingTransformer implements CachedClasspathTransformer.Transform {
                 bootstrapMethodHandle.getOwner().equals(GROOVY_INDY_INTERFACE_V7_TYPE)) &&
                 bootstrapMethodHandle.getName().equals("bootstrap") &&
                 bootstrapMethodHandle.getDesc().equals(GROOVY_INDY_INTERFACE_BOOTSTRAP_METHOD_DESCRIPTOR);
-        }
-    }
-
-    private static class LambdaFactoryDetails {
-        final String name;
-        final String descriptor;
-        final Handle bootstrapMethodHandle;
-        final List<?> bootstrapMethodArguments;
-
-        public LambdaFactoryDetails(String name, String descriptor, Handle bootstrapMethodHandle, List<?> bootstrapMethodArguments) {
-            this.name = name;
-            this.descriptor = descriptor;
-            this.bootstrapMethodHandle = bootstrapMethodHandle;
-            this.bootstrapMethodArguments = bootstrapMethodArguments;
         }
     }
 }
